pyproject.toml
[tool.poetry]
name = "wichtel"
version = "0.1.0"
description = "A random match generator for christmas presents"
authors = ["Olaf Gladis"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
httpx = "^0.20.0"
pydantic = {extras = ["dotenv"], version = "^1.8.2"}
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
pytest-watch = "^4.2.0"
pytest = "^6.2.5"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
